Reduce component size
